refactor(aichat): migrate to BlocScreen pattern
Mirrors the convention introduced in #198 for cook and bottomNavOrder:

- AiChatScreen and AiChatPreviews move from aichat/public to
  aichat/impl/ui. aichat/public retains the BLoC interface, model, test
  tags, and TextData.
- AiChatBloc extends BlocScreen; AiChatBlocImpl overrides Content() to
  call AiChatScreen(this). Preview Blocs do the same so IDE previews
  and snapshot tests keep working.
- RootBloc.Child.AiChat gains BlocScreen by bloc; RootScreen drops the
  AiChat special-case and renders via activeChild.instance.Content().
- aichat/public sets publicResClass = true so impl can read Res.
- aichat/impl adds the compose plugin and pulls in the markdown
  renderer + Compose deps it now hosts directly.
- Snapshot tests depend on aichat:impl for the moved previews.

No behavior change; recorded PNGs unchanged.
