+++ client/aichat/impl/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     alias(libs.plugins.kmpLibrary)
+    alias(libs.plugins.compose)
     alias(libs.plugins.kotlinSerialization)
 }
 
@@ -9,11 +10,16 @@ kotlin {
             implementation(projects.client.aichat.public)
             implementation(projects.client.shared)
             implementation(projects.client.database.core)
+            implementation(projects.client.text.public)
+            implementation(projects.client.ui.public)
             implementation(libs.arkivanov.decompose.core)
+            implementation(libs.arkivanov.decompose.compose.extensions)
             implementation(libs.kotlinx.datetime)
             implementation(libs.kotlinx.serialization.json)
             implementation(libs.ktor.client.content.negotiation)
             implementation(libs.ktor.serialization.json)
+            implementation(libs.multiplatform.markdown.renderer.m3)
+            implementation(compose.components.resources)
         }
         commonTest.dependencies { implementation(libs.kotlin.coroutines.test) }
         getByName("androidMain").dependencies { implementation(libs.ktor.client.cio) }
