+++ client/ui/screenshot-test/build.gradle.kts
@@ -35,6 +35,7 @@ dependencies {
     implementation(project(":client:ui:public"))
     implementation(project(":client:text:public"))
     implementation(project(":client:aichat:public"))
+    implementation(project(":client:aichat:impl"))
     implementation(project(":client:auth:ui:public"))
     implementation(project(":client:auth:ui:impl"))
     implementation(project(":client:bottomnav:public"))
