+++ client/ui/screenshot-test/src/screenshotTest/kotlin/com/plusmobileapps/chefmate/ui/screenshot/AiChatScreenshotTest.kt
@@ -9,18 +9,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tools.screenshot.PreviewTest
 import com.plusmobileapps.chefmate.aichat.AiChatBloc
-import com.plusmobileapps.chefmate.aichat.AiChatScreen
-import com.plusmobileapps.chefmate.aichat.previewAiChatBloc
-import com.plusmobileapps.chefmate.aichat.previewAiChatBlocEmpty
-import com.plusmobileapps.chefmate.aichat.previewAiChatBlocError
-import com.plusmobileapps.chefmate.aichat.previewAiChatBlocStreaming
+import com.plusmobileapps.chefmate.aichat.impl.ui.previewAiChatBloc
+import com.plusmobileapps.chefmate.aichat.impl.ui.previewAiChatBlocEmpty
+import com.plusmobileapps.chefmate.aichat.impl.ui.previewAiChatBlocError
+import com.plusmobileapps.chefmate.aichat.impl.ui.previewAiChatBlocStreaming
 import com.plusmobileapps.chefmate.ui.theme.ChefMateTheme
 
 @Composable
 private fun AiChatScreenshot(bloc: AiChatBloc, darkTheme: Boolean = false) {
     ChefMateTheme(darkTheme = darkTheme) {
         Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background) {
-            AiChatScreen(bloc = bloc)
+            bloc.Content(Modifier)
         }
     }
 }
